+++ default_www/backend/modules/analytics/widgets/traffic_sources.php
@@ -108,16 +108,16 @@ class BackendAnalyticsWidgetTrafficSources extends BackendBaseWidget
 		if(!empty($results))
 		{
 			// get the datagrid
-			$datagrid = new BackendDataGridArray($results);
+			$dataGrid = new BackendDataGridArray($results);
 
 			// no pagination
-			$datagrid->setPaging(false);
+			$dataGrid->setPaging(false);
 
 			// hide columns
-			$datagrid->setColumnsHidden('id', 'date');
+			$dataGrid->setColumnsHidden('id', 'date');
 
 			// parse the datagrid
-			$this->tpl->assign('dgAnalyticsKeywords', $datagrid->getContent());
+			$this->tpl->assign('dgAnalyticsKeywords', $dataGrid->getContent());
 		}
 
 		// get date
@@ -143,19 +143,19 @@ class BackendAnalyticsWidgetTrafficSources extends BackendBaseWidget
 		if(!empty($results))
 		{
 			// get the datagrid
-			$datagrid = new BackendDataGridArray($results);
+			$dataGrid = new BackendDataGridArray($results);
 
 			// no pagination
-			$datagrid->setPaging(false);
+			$dataGrid->setPaging(false);
 
 			// hide columns
-			$datagrid->setColumnsHidden('id', 'date');
+			$dataGrid->setColumnsHidden('id', 'date');
 
 			// set url
-			$datagrid->setColumnURL('referrer', 'http://[referrer]');
+			$dataGrid->setColumnURL('referrer', 'http://[referrer]');
 
 			// parse the datagrid
-			$this->tpl->assign('dgAnalyticsReferrers', $datagrid->getContent());
+			$this->tpl->assign('dgAnalyticsReferrers', $dataGrid->getContent());
 		}
 	}
 }