+++ default_www/backend/modules/blog/actions/edit.php
@@ -23,9 +23,9 @@ class BackendBlogEdit extends BackendBaseActionEdit
 
 
 	/**
-	 * Datagrid for the drafts
+	 * DataGrid for the drafts
 	 *
-	 * @var	BackendDatagrid
+	 * @var	BackendDataGrid
 	 */
 	private $dgDrafts;