+++ default_www/backend/modules/groups/actions/add.php
@@ -331,6 +331,8 @@ class BackendGroupsAdd extends BackendBaseActionAdd
 		// init vars
 		$modulesDenied = array();
 		$modulesGranted = array();
+		$actionsDenied = array();
+		$actionsGranted = array();
 		$checkedModules = array();
 		$uncheckedModules = array();
 
@@ -341,8 +343,8 @@ class BackendGroupsAdd extends BackendBaseActionAdd
 			$bits = explode('_', $permission->getName());
 
 			// convert camelcasing to underscore notation
-			$module = strtolower(preg_replace('/([A-Z])/', '_${1}', lcfirst($bits[1])));
-			$action = strtolower(preg_replace('/([A-Z])/', '_${1}', lcfirst($bits[2])));
+			$module = trim(strtolower(preg_replace('/([A-Z])/', '_${1}', $bits[1])), '_');
+			$action = trim(strtolower(preg_replace('/([A-Z])/', '_${1}', $bits[2])), '_');
 
 			// permission checked?
 			if($permission->getChecked())
@@ -372,8 +374,8 @@ class BackendGroupsAdd extends BackendBaseActionAdd
 			$bits = explode('_', $permission->getName());
 
 			// convert camelcasing to underscore notation
-			$module = strtolower(preg_replace('/([A-Z])/', '_${1}', lcfirst($bits[1])));
-			$group = strtolower(preg_replace('/([A-Z])/', '_${1}', lcfirst($bits[3])));
+			$module = trim(strtolower(preg_replace('/([A-Z])/', '_${1}', $bits[1])), '_');
+			$group = trim(strtolower(preg_replace('/([A-Z])/', '_${1}', $bits[3])), '_');
 
 			// create new item
 			$moduleItem = array('group_id' => $this->id, 'module' => $module);
@@ -454,7 +456,7 @@ class BackendGroupsAdd extends BackendBaseActionAdd
 				if($preset->getChecked())
 				{
 					// convert camelcasing to underscore notation
-					$selected = strtolower(preg_replace('/([A-Z])/', '_${1}', lcfirst(str_replace('widgets_', '', $preset->getName()))));
+					$selected = trim(strtolower(preg_replace('/([A-Z])/', '_${1}', str_replace('widgets_', '', $preset->getName()))), '_');
 
 					// if right widget set visible
 					if($selected === $widget['widget']) $this->dashboardSequence[$widget['module']][$widget['widget']]['present'] = true;
@@ -556,7 +558,7 @@ class BackendGroupsAdd extends BackendBaseActionAdd
 			$actionGrid->setPaging(false);
 
 			// get content of datagrids
-			$permissionBoxes[$key]['actions']['datagrid'] = $actionGrid->getContent();
+			$permissionBoxes[$key]['actions']['dataGrid'] = $actionGrid->getContent();
 			$permissionBoxes[$key]['chk'] = $this->frm->addCheckbox($module['label'], null, 'inputCheckbox checkBeforeUnload selectAll')->parse();
 		}