+++ default_www/backend/modules/mailmotor/actions/mass_campaign_action.php
@@ -38,7 +38,7 @@ class BackendMailmotorMassCampaignAction extends BackendBaseAction
 		}
 
 		// redirect
-		$this->redirect(BackendModel::createURLForAction('campaigns') . '&report=delete_campaigns');
+		$this->redirect(BackendModel::createURLForAction('campaigns') . '&report=delete-campaigns');
 	}
 }