+++ default_www/backend/modules/mailmotor/widgets/statistics.php
@@ -90,13 +90,13 @@ class BackendMailmotorWidgetStatistics extends BackendBaseWidget
 		if(!empty($results))
 		{
 			// get the datagrid
-			$datagrid = new BackendDataGridArray($results);
+			$dataGrid = new BackendDataGridArray($results);
 
 			// no pagination
-			$datagrid->setPaging(false);
+			$dataGrid->setPaging(false);
 
 			// parse the datagrid
-			$this->tpl->assign('dgMailmotorStatistics', $datagrid->getContent());
+			$this->tpl->assign('dgMailmotorStatistics', $dataGrid->getContent());
 		}
 	}
 
@@ -115,19 +115,19 @@ class BackendMailmotorWidgetStatistics extends BackendBaseWidget
 		if(!empty($results))
 		{
 			// get the datagrid
-			$datagrid = new BackendDataGridArray($results);
+			$dataGrid = new BackendDataGridArray($results);
 
 			// no pagination
-			$datagrid->setPaging(false);
+			$dataGrid->setPaging(false);
 
 			// set edit link
-			$datagrid->setColumnURL('email', BackendModel::createURLForAction('edit_address', 'mailmotor') . '&amp;email=[email]');
+			$dataGrid->setColumnURL('email', BackendModel::createURLForAction('edit_address', 'mailmotor') . '&amp;email=[email]');
 
 			// set column functions
-			$datagrid->setColumnFunction(array('BackendDatagridFunctions', 'getTimeAgo'), array('[created_on]'), 'created_on', true);
+			$dataGrid->setColumnFunction(array('BackendDataGridFunctions', 'getTimeAgo'), array('[created_on]'), 'created_on', true);
 
 			// parse the datagrid
-			$this->tpl->assign('dgMailmotorSubscriptions', $datagrid->getContent());
+			$this->tpl->assign('dgMailmotorSubscriptions', $dataGrid->getContent());
 		}
 	}
 
@@ -146,19 +146,19 @@ class BackendMailmotorWidgetStatistics extends BackendBaseWidget
 		if(!empty($results))
 		{
 			// get the datagrid
-			$datagrid = new BackendDataGridArray($results);
+			$dataGrid = new BackendDataGridArray($results);
 
 			// no pagination
-			$datagrid->setPaging(false);
+			$dataGrid->setPaging(false);
 
 			// set edit link
-			$datagrid->setColumnURL('email', BackendModel::createURLForAction('edit_address', 'mailmotor') . '&amp;email=[email]');
+			$dataGrid->setColumnURL('email', BackendModel::createURLForAction('edit_address', 'mailmotor') . '&amp;email=[email]');
 
 			// set column functions
-			$datagrid->setColumnFunction(array('BackendDatagridFunctions', 'getTimeAgo'), array('[created_on]'), 'created_on', true);
+			$dataGrid->setColumnFunction(array('BackendDataGridFunctions', 'getTimeAgo'), array('[created_on]'), 'created_on', true);
 
 			// parse the datagrid
-			$this->tpl->assign('dgMailmotorUnsubscriptions', $datagrid->getContent());
+			$this->tpl->assign('dgMailmotorUnsubscriptions', $dataGrid->getContent());
 		}
 	}