+++ default_www/backend/modules/settings/actions/index.php
@@ -52,7 +52,7 @@ class BackendSettingsIndex extends BackendBaseActionIndex
 		// validate the form
 		$this->validateForm();
 
-		// parse the datagrid
+		// parse
 		$this->parse();
 
 		// display the page