+++ default_www/frontend/core/engine/base.php
@@ -352,24 +352,6 @@ class FrontendBaseBlock
 	/**
 	 * Add a javascript file into the array
 	 *
-	 * @deprecated	Deprecated since version 2.2.0. Will be removed in the next version.
-	 *
-	 * @return	void
-	 * @param 	string $file						The path to the javascript-file that should be loaded.
-	 * @param 	bool[optional] $overwritePath		Whether or not to add the module to this path. Module path is added by default.
-	 * @param	bool[optional] $minify				Should the file be minified?
-	 * @param	bool[optional] $parseThroughPHP		Should the file be parsed through PHP?
-	 * @param	bool[optional] $addTimestamp		May we add a timestamp for caching purposes?
-	 */
-	public function addJavascript($file, $overwritePath = false, $minify = true, $parseThroughPHP = false, $addTimestamp = null)
-	{
-		$this->addJS($file, $overwritePath, $minify, $parseThroughPHP, $addTimestamp);
-	}
-
-
-	/**
-	 * Add a javascript file into the array
-	 *
 	 * @return	void
 	 * @param 	string $file						The path to the javascript-file that should be loaded.
 	 * @param 	bool[optional] $overwritePath		Whether or not to add the module to this path. Module path is added by default.
@@ -851,23 +833,6 @@ class FrontendBaseWidget
 
 		// add css to the header
 		$this->header->addCSS($file, $minify, $addTimestamp);
-	}
-
-
-	/**
-	 * Add a javascript file into the array
-	 *
-	 * @deprecated	Deprecated since version 2.2.0. Will be removed in the next version.
-	 *
-	 * @return	void
-	 * @param 	string $file						The path to the javascript-file that should be loaded.
-	 * @param 	bool[optional] $overwritePath		Whether or not to add the module to this path. Module path is added by default.
-	 * @param	bool[optional] $minify				Should the file be minified?
-	 * @param	bool[optional] $parseThroughPHP		Should the file be parsed through PHP?
-	 */
-	public function addJavascript($file, $overwritePath = false, $minify = true, $parseThroughPHP = false)
-	{
-		$this->addJS($file, $overwritePath, $minify, $parseThroughPHP);
 	}